Split javascript export template into multiple files
